fix eslint rule, pull visualizationtracker logic out into hooks file
