+++ context/.eslintrc.yml
@@ -49,13 +49,13 @@ rules:
   import/prefer-default-export: [0] # Hit eslint error: SyntaxError: Unexpected token, expected {
   react/jsx-props-no-spreading: [0] # common pattern, but with TypeScript it"s pretty safe
   no-restricted-syntax: # Disallow certain syntax forms while still allowing `for..of` loops
-    - error
+    - 2
     - selector: ForInStatement
       message: for..in loops iterate over the entire prototype chain, which is virtually never what you want. Use Object.{keys,values,entries}, and iterate over the resulting array.
     - selector: LabeledStatement
       message: Labels are a form of GOTO; using them makes code confusing and hard to maintain and understand.
     - selector: WithStatement
-      message: `with` is disallowed in strict mode because it makes code impossible to predict and optimize.
+      message: with is disallowed in strict mode because it makes code impossible to predict and optimize.
   jsx-a11y/label-has-associated-control:
     - 2
     - controlComponents:
